Seperate PokemonType & PokemonChoice header
