+++ PokemonType.hpp
@@ -1,0 +1,9 @@
+#pragma once
+// Define an enum for Pokemon types
+enum class PokemonType {
+    FIRE,
+    GRASS,
+    WATER,
+    ELECTRIC,
+    NORMAL // Added for the default constructor
+};
