+++ main/PokemonChoice.hpp
@@ -1,0 +1,8 @@
+#pragma once
+// Define an enum for Pokemon choices
+enum class PokemonChoice {
+    CHARMANDER = 1,
+    BULBASAUR,
+    SQUIRTLE,
+    PIKACHU // Default choice
+};
